requirements/base.txt
Django
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.0
django-cors-headers==4.3.1
python-decouple==3.8
psycopg2-binary==2.9.7
Pillow==10.3.0
celery==5.3.1
redis==4.6.0
requests
stripe
paypal-checkout-serversdk==1.0.1
googlemaps==4.10.0
openai==0.28.1
google-generativeai==0.5.0
openpyxl==3.1.2
python-docx==0.8.11
reportlab==4.0.4
django-extensions==3.2.3
pandas
openai